Add files via upload
The company “Home SmartApps” has decided to design a new intelligent
refrigerator. Besides the hardware, they need a software application to
manage the refrigerator. Each Type of food has a name, a category (may
be dairy, sweets, meat or fruit), a quantity and an expiration date.
a. The application must allow adding, deleting and updating a type of
food. A type of food is uniquely identified by name and category. If a type
of food that already exists is added, its quantity will be updated (the new
quantity is added to the existing one).
b. The application should offer the possibility to display all the types
of food whose names contain a given string (if the string is empty, all types of food from the
refrigerator are considered) and they will be shown sorted ascending by their quantities.
c. The application should be able to display all types of food of a given category (if the category is
empty, all types of food will be considered) whose expiration dates are close (exp